Update sidebar and config structure
