+++ docs/.vuepress/config/versions/1.x.x.js
@@ -1,0 +1,37 @@
+module.exports = [
+  {
+    collapsable: false,
+    title: 'UsefulLinks',
+    children: [
+      ['/1.x.x/', 'Introduction'],
+      ['https://strapi.io', 'Strapi Website'],
+      ['https://github.com/strapi/strapi', 'GitHub Repository'],
+      [
+        'https://github.com/strapi/strapi/blob/master/CONTRIBUTING.md',
+        'Contribution Guide',
+      ],
+    ],
+  },
+  '/1.x.x/admin.md',
+  '/1.x.x/configuration.md',
+  '/1.x.x/email.md',
+  '/1.x.x/introduction.md',
+  '/1.x.x/queries.md',
+  '/1.x.x/response.md',
+  '/1.x.x/sessions.md',
+  '/1.x.x/testing.md',
+  '/1.x.x/views.md',
+  '/1.x.x/blueprints.md',
+  '/1.x.x/context.md',
+  '/1.x.x/graphql.md',
+  '/1.x.x/logging.md',
+  '/1.x.x/router.md',
+  '/1.x.x/upload.md',
+  '/1.x.x/cli.md',
+  '/1.x.x/customization.md',
+  '/1.x.x/internationalization.md',
+  '/1.x.x/models.md',
+  '/1.x.x/request.md',
+  '/1.x.x/services.md',
+  '/1.x.x/users.md',
+];
