+++ docs/.vuepress/config/versions/3.0.0-beta.x.js
@@ -1,0 +1,84 @@
+module.exports = [
+  {
+    collapsable: false,
+    title: '🚀 Getting started',
+    children: [
+      '/3.0.0-beta.x/getting-started/introduction',
+      '/3.0.0-beta.x/getting-started/install-requirements',
+      '/3.0.0-beta.x/getting-started/quick-start',
+      '/3.0.0-beta.x/getting-started/quick-start-tutorial',
+    ],
+  },
+  {
+    collapsable: true,
+    title: '💡 Guides',
+    children: [
+      '/3.0.0-beta.x/concepts/concepts',
+      '/3.0.0-beta.x/guides/api-documentation',
+      '/3.0.0-beta.x/guides/authentication',
+      '/3.0.0-beta.x/configurations/configurations',
+      '/3.0.0-beta.x/guides/controllers',
+      '/3.0.0-beta.x/guides/databases',
+      '/3.0.0-beta.x/guides/deployment',
+      '/3.0.0-beta.x/guides/email',
+      '/3.0.0-beta.x/guides/upload',
+      '/3.0.0-beta.x/guides/filters',
+      '/3.0.0-beta.x/guides/graphql',
+      '/3.0.0-beta.x/guides/i18n',
+      '/3.0.0-beta.x/guides/models',
+      '/3.0.0-beta.x/guides/policies',
+      '/3.0.0-beta.x/guides/public-assets',
+      '/3.0.0-beta.x/guides/requests',
+      '/3.0.0-beta.x/guides/responses',
+      '/3.0.0-beta.x/guides/routing',
+      '/3.0.0-beta.x/guides/services',
+      '/3.0.0-beta.x/guides/queries',
+      '/3.0.0-beta.x/guides/webhooks',
+    ],
+  },
+  {
+    collapsable: true,
+    title: '⚙️️ Advanced',
+    children: [
+      '/3.0.0-beta.x/advanced/customize-admin',
+      '/3.0.0-beta.x/advanced/hooks',
+      '/3.0.0-beta.x/advanced/logging',
+      '/3.0.0-beta.x/advanced/middlewares',
+      '/3.0.0-beta.x/advanced/usage-information',
+    ],
+  },
+  {
+    collapsable: true,
+    title: '🔌 Local plugins',
+    children: [
+      '/3.0.0-beta.x/plugin-development/quick-start',
+      '/3.0.0-beta.x/plugin-development/plugin-architecture',
+      '/3.0.0-beta.x/plugin-development/backend-development',
+      '/3.0.0-beta.x/plugin-development/frontend-development',
+
+      // '/3.0.0-beta.x/plugin-development/ui-components', TODO: Add this file
+    ],
+  },
+  {
+    collapsable: true,
+    title: '💻 Command Line Interface',
+    children: ['/3.0.0-beta.x/cli/CLI'],
+  },
+  {
+    collapsable: true,
+    title: '🏗 API Reference',
+    children: ['/3.0.0-beta.x/api-reference/reference'],
+  },
+  {
+    collapsable: false,
+    title: '📚 Resources',
+    children: [
+      [
+        'https://github.com/strapi/strapi/blob/master/CONTRIBUTING.md',
+        'Contributing guide',
+      ],
+      '/3.0.0-beta.x/migration-guide/',
+      '/3.0.0-beta.x/articles/',
+    ],
+  },
+]
