+++ docs/.vuepress/theme/SidebarLink.vue
@@ -9,11 +9,11 @@ export default {
   render (h, { parent: { $page, $site, $route }, props: { item, open, collapsable }}) {
     // use custom active class matching logic
     // due to edge case of paths ending with / + hash
-    const selfActive = isActive($route, item.path)
+    const selfActive = isActive($route, item.path, $page.frontmatter)
     // for sidebar: auto pages, a hash link should be active if one of its child
     // matches
     const active = item.type === 'auto'
-      ? selfActive || item.children.some(c => isActive($route, item.basePath + '#' + c.slug))
+      ? selfActive || item.children.some(c => isActive($route, item.basePath + '#' + c.slug, $page.frontmatter))
       : selfActive
     const hidden = $site.themeConfig.hiddenLinks.includes(item.path);
     const link = renderLink(h, item.path, item.title || item.path, active, hidden)
@@ -22,13 +22,13 @@ export default {
       : $site.themeConfig.sidebarDepth
     const maxDepth = configDepth == null ? 1 : configDepth
     if (item.type === 'auto') {
-      return [link, renderChildren(h, item.children, item.basePath, $route, maxDepth)]
+      return [link, renderChildren(h, item.children, item.basePath, $route, maxDepth, $page.frontmatter)]
     } else if (active && item.headers && !hashRE.test(item.path)) {
       const children = groupHeaders(item.headers)
-      return [link, renderChildren(h, children, item.path, $route, maxDepth)]
+      return [link, renderChildren(h, children, item.path, $route, maxDepth, $page.frontmatter)]
     } else if (collapsable && open && hidden && !active && item.headers && !hashRE.test(item.path)) {
       const children = groupHeaders(item.headers)
-      return [link, renderChildren(h, children, item.path, $route, maxDepth)]
+      return [link, renderChildren(h, children, item.path, $route, maxDepth, $page.frontmatter)]
     } else {
       return link
     }
@@ -58,13 +58,13 @@ function renderLink (h, to, text, active, hidden) {
   }, text)
 }
 
-function renderChildren (h, children, path, route, maxDepth, depth = 1) {
+function renderChildren (h, children, path, route, maxDepth, frontmatter, depth = 1) {
   if (!children || depth > maxDepth) return null
   return h('ul', { class: 'sidebar-sub-headers' }, children.map(c => {
-    const active = isActive(route, path + '#' + c.slug)
+    const active = isActive(route, path + '#' + c.slug, frontmatter)
     return h('li', { class: 'sidebar-sub-header' }, [
       renderLink(h, path + '#' + c.slug, c.title, active),
-      renderChildren(h, c.children, path, route, maxDepth, depth + 1)
+      renderChildren(h, c.children, path, route, maxDepth, frontmatter, depth + 1)
     ])
   }))
 }
