+++ docs/package.json
@@ -1,18 +1,13 @@
 {
-  "name": "docs",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
   "scripts": {
     "dev": "vuepress dev",
     "build": "vuepress build"
   },
-  "author": "",
-  "license": "ISC",
   "dependencies": {
     "cache-loader": "1.2.2",
     "directory-tree": "^2.1.0",
     "markdown-it-decorate": "^1.2.2",
+    "markdown-it-include": "^1.1.0",
     "vue-tabs-component": "^1.5.0",
     "vuepress": "^0.14.2",
     "vuepress-tabs": "^0.1.6"
